Update bundled openresty/luajit2 to v2.1-20200102
Update the version of openresty/luajit2 in the 1.15.8.x branch to
v2.1-20200102.

This is to allow a release of openresty that includes a version of
luajit that supports the s390x architecture to be created. Support for
s390x was added to openresty/luajit2 by this PR -

https://github.com/openresty/luajit2/pull/50

which is included in openresty/luajit2 release v2.1-20200102, however
the bundled release of openresty/luajit2 in the latest release of
openresty, version 1.15.8.3, is v2.1-20190507, which has no support
for s390x.

Commit added by Matthew Arnold <mattarno@uk.ibm.com>